+++ input-stream/src/main/java/software/amazon/s3/analyticsaccelerator/io/physical/data/BlockManager.java
@@ -15,57 +15,55 @@
  */
 package software.amazon.s3.analyticsaccelerator.io.physical.data;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.io.Closeable;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalLong;
+import java.util.concurrent.ExecutorService;
 import lombok.NonNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import software.amazon.s3.analyticsaccelerator.common.Metrics;
 import software.amazon.s3.analyticsaccelerator.common.Preconditions;
-import software.amazon.s3.analyticsaccelerator.common.telemetry.Operation;
 import software.amazon.s3.analyticsaccelerator.common.telemetry.Telemetry;
 import software.amazon.s3.analyticsaccelerator.io.physical.PhysicalIOConfiguration;
-import software.amazon.s3.analyticsaccelerator.io.physical.prefetcher.SequentialPatternDetector;
 import software.amazon.s3.analyticsaccelerator.io.physical.prefetcher.SequentialReadProgression;
-import software.amazon.s3.analyticsaccelerator.request.ObjectClient;
-import software.amazon.s3.analyticsaccelerator.request.ObjectMetadata;
-import software.amazon.s3.analyticsaccelerator.request.Range;
-import software.amazon.s3.analyticsaccelerator.request.ReadMode;
-import software.amazon.s3.analyticsaccelerator.util.*;
+import software.amazon.s3.analyticsaccelerator.io.physical.reader.StreamReader;
+import software.amazon.s3.analyticsaccelerator.request.*;
+import software.amazon.s3.analyticsaccelerator.util.BlockKey;
+import software.amazon.s3.analyticsaccelerator.util.ObjectKey;
+import software.amazon.s3.analyticsaccelerator.util.OpenStreamInformation;
 
 /** Implements a Block Manager responsible for planning and scheduling reads on a key. */
 public class BlockManager implements Closeable {
   private final ObjectKey objectKey;
   private final ObjectMetadata metadata;
-  private final BlockStore blockStore;
-  private final ObjectClient objectClient;
+
+  @SuppressFBWarnings(
+      value = "URF_UNREAD_FIELD",
+      justification = "Field is injected and may be used in the future")
   private final Telemetry telemetry;
-  private final SequentialPatternDetector patternDetector;
-  private final SequentialReadProgression sequentialReadProgression;
-  private final IOPlanner ioPlanner;
+
   private final PhysicalIOConfiguration configuration;
-  private final RangeOptimiser rangeOptimiser;
-  private OpenStreamInformation openStreamInformation;
   private final Metrics aggregatingMetrics;
   private final BlobStoreIndexCache indexCache;
-  private static final String OPERATION_MAKE_RANGE_AVAILABLE = "block.manager.make.range.available";
-
-  private static final Logger LOG = LoggerFactory.getLogger(BlockManager.class);
+  private final StreamReader streamReader;
+  private final BlockStore blockStore;
+  private final SequentialReadProgression sequentialReadProgression;
+  private final RangeOptimiser rangeOptimiser;
 
   /**
    * Constructs a new BlockManager.
    *
-   * @param objectKey the etag and S3 URI of the object
-   * @param objectClient object client capable of interacting with the underlying object store
-   * @param telemetry an instance of {@link Telemetry} to use
-   * @param metadata the metadata for the object
-   * @param configuration the physicalIO configuration
-   * @param aggregatingMetrics factory metrics
-   * @param indexCache blobstore index cache
+   * @param objectKey the key representing the S3 object, including its URI and ETag
+   * @param objectClient the client used to fetch object content from S3
+   * @param metadata metadata associated with the S3 object, including content length
+   * @param telemetry the telemetry interface used for logging or instrumentation
+   * @param configuration configuration for physical IO operations (e.g., read buffer size)
+   * @param aggregatingMetrics the metrics aggregator for performance or usage monitoring
+   * @param indexCache cache for blob index metadata (if applicable)
    * @param openStreamInformation contains stream information
+   * @param threadPool Thread pool
    */
   public BlockManager(
       @NonNull ObjectKey objectKey,
@@ -75,51 +73,20 @@ public class BlockManager implements Closeable {
       @NonNull PhysicalIOConfiguration configuration,
       @NonNull Metrics aggregatingMetrics,
       @NonNull BlobStoreIndexCache indexCache,
-      @NonNull OpenStreamInformation openStreamInformation) {
+      @NonNull OpenStreamInformation openStreamInformation,
+      @NonNull ExecutorService threadPool) {
     this.objectKey = objectKey;
-    this.objectClient = objectClient;
     this.metadata = metadata;
     this.telemetry = telemetry;
     this.configuration = configuration;
     this.aggregatingMetrics = aggregatingMetrics;
     this.indexCache = indexCache;
-    this.blockStore = new BlockStore(objectKey, metadata, aggregatingMetrics, indexCache);
-    this.patternDetector = new SequentialPatternDetector(blockStore);
+    this.blockStore = new BlockStore(indexCache, aggregatingMetrics, configuration);
+    this.streamReader =
+        new StreamReader(
+            objectClient, objectKey, threadPool, this::removeBlocks, openStreamInformation);
     this.sequentialReadProgression = new SequentialReadProgression(configuration);
-    this.ioPlanner = new IOPlanner(blockStore);
     this.rangeOptimiser = new RangeOptimiser(configuration);
-    this.openStreamInformation = openStreamInformation;
-
-    prefetchSmallObject();
-  }
-
-  /**
-   * Initializes the BlockManager with small object prefetching if applicable. This is done
-   * asynchronously to avoid blocking the constructor.
-   */
-  private void prefetchSmallObject() {
-    if (AnalyticsAcceleratorUtils.isSmallObject(configuration, metadata.getContentLength())) {
-      try {
-        makeRangeAvailable(0, metadata.getContentLength(), ReadMode.SMALL_OBJECT_PREFETCH);
-      } catch (IOException e) {
-        LOG.debug("Failed to prefetch small object for key: {}", objectKey.getS3URI().getKey(), e);
-      }
-    }
-  }
-
-  /** @return true if blockstore is empty */
-  public boolean isBlockStoreEmpty() {
-    return blockStore.isBlockStoreEmpty();
-  }
-
-  /**
-   * Given the position of a byte, return the block holding it.
-   *
-   * @param pos the position of a byte
-   * @return the Block holding the byte or empty if the byte is not in the BlockStore
-   */
-  public synchronized Optional<Block> getBlock(long pos) {
-    return this.blockStore.getBlock(pos);
   }
 
   /**
@@ -127,32 +94,10 @@ public class BlockManager implements Closeable {
    *
    * @param pos the position of the byte
    * @param readMode whether this ask corresponds to a sync or async read
-   * @throws IOException if an I/O error occurs
    */
   public synchronized void makePositionAvailable(long pos, ReadMode readMode) throws IOException {
     Preconditions.checkArgument(0 <= pos, "`pos` must not be negative");
-
-    // Position is already available --> return corresponding block
-    if (getBlock(pos).isPresent()) {
-      return;
-    }
-
     makeRangeAvailable(pos, 1, readMode);
-  }
-
-  private boolean isRangeAvailable(long pos, long len) throws IOException {
-    Preconditions.checkArgument(0 <= pos, "`pos` must not be negative");
-    Preconditions.checkArgument(0 <= len, "`len` must not be negative");
-
-    long lastByteOfRange = pos + len - 1;
-
-    OptionalLong nextMissingByte = blockStore.findNextMissingByte(pos);
-    if (nextMissingByte.isPresent()) {
-      return lastByteOfRange < nextMissingByte.getAsLong();
-    }
-
-    // If there is no missing byte after pos, then the whole object is already fetched
-    return true;
   }
 
   /**
@@ -163,79 +108,94 @@ public class BlockManager implements Closeable {
    * @param pos start of a read
    * @param len length of the read
    * @param readMode whether this ask corresponds to a sync or async read
-   * @throws IOException if an I/O error occurs
    */
   public synchronized void makeRangeAvailable(long pos, long len, ReadMode readMode)
       throws IOException {
     Preconditions.checkArgument(0 <= pos, "`pos` must not be negative");
     Preconditions.checkArgument(0 <= len, "`len` must not be negative");
 
-    if (isRangeAvailable(pos, len)) {
-      return;
+    long endPos = pos + len - 1;
+
+    // Range is available, return
+    if (isRangeAvailable(pos, endPos)) return;
+
+    long generation = getGeneration(pos, readMode);
+
+    /*
+     There are three different range length we need to consider.
+     1/ Length of the requested read
+     2/ Read ahead bytes length
+     3/ Sequential read pattern length
+     We need to send the request for the largest of one of these 3 lengths
+     to find the optimum request length
+    */
+    long maxReadLength =
+        Math.max(
+            Math.max(len, configuration.getReadAheadBytes()),
+            sequentialReadProgression.getSizeForGeneration(generation));
+    long effectiveEnd = truncatePos(pos + maxReadLength - 1);
+
+    // Find missing blocks for given range.
+    // measure is false because we already add statistics in isRangeAvailable(),
+    // so no need to add measure
+    List<Integer> missingBlockIndexes =
+        blockStore.getMissingBlockIndexesInRange(pos, effectiveEnd, false);
+
+    // Return if all blocks are in store
+    if (missingBlockIndexes.isEmpty()) return;
+
+    // Split missing blocks into groups of sequential indexes that respect maximum range size
+    List<List<Integer>> groupedReads = splitReads(missingBlockIndexes);
+
+    // Process each group separately to optimize read operations
+    for (List<Integer> group : groupedReads) {
+      // Create blocks for this group of sequential indexes
+      List<Block> blocksToFill = new ArrayList<>();
+      for (int blockIndex : group) {
+        BlockKey blockKey = new BlockKey(objectKey, getBlockIndexRange(blockIndex));
+        Block block =
+            new Block(
+                blockKey,
+                generation,
+                this.indexCache,
+                this.aggregatingMetrics,
+                this.configuration.getBlockReadTimeout());
+        // Add block to the store for future reference
+        blockStore.add(block);
+        blocksToFill.add(block);
+      }
+
+      // Perform a single read operation for this group of sequential blocks
+      streamReader.read(blocksToFill, readMode);
     }
-
-    // In case of a sequential reading pattern, calculate the generation and adjust the requested
-    // effectiveEnd of the requested range
-    long effectiveEnd = pos + Math.max(len, configuration.getReadAheadBytes()) - 1;
-
-    // Check sequential prefetching. If read mode is ASYNC, that is the request is from the parquet
-    // prefetch path, then do not extend the request.
-    // TODO: Improve readModes, as tracked in
-    // https://github.com/awslabs/analytics-accelerator-s3/issues/195
-    final long generation;
-    if (readMode != ReadMode.ASYNC && patternDetector.isSequentialRead(pos)) {
-      generation = patternDetector.getGeneration(pos);
-      effectiveEnd =
-          Math.max(
-              effectiveEnd,
-              truncatePos(pos + sequentialReadProgression.getSizeForGeneration(generation)));
-    } else {
-      generation = 0;
-    }
-
-    // Fix "effectiveEnd", so we can pass it into the lambda
-    final long effectiveEndFinal = effectiveEnd;
-    this.telemetry.measureStandard(
-        () ->
-            Operation.builder()
-                .name(OPERATION_MAKE_RANGE_AVAILABLE)
-                .attribute(StreamAttributes.uri(this.objectKey.getS3URI()))
-                .attribute(StreamAttributes.etag(this.objectKey.getEtag()))
-                .attribute(StreamAttributes.range(pos, pos + len - 1))
-                .attribute(StreamAttributes.effectiveRange(pos, effectiveEndFinal))
-                .attribute(StreamAttributes.generation(generation))
-                .build(),
-        () -> {
-          // Determine the missing ranges and fetch them
-          List<Range> missingRanges =
-              ioPlanner.planRead(pos, effectiveEndFinal, getLastObjectByte());
-          List<Range> splits = rangeOptimiser.splitRanges(missingRanges);
-          for (Range r : splits) {
-            BlockKey blockKey = new BlockKey(objectKey, r);
-            Block block =
-                new Block(
-                    blockKey,
-                    objectClient,
-                    telemetry,
-                    generation,
-                    readMode,
-                    this.configuration.getBlockReadTimeout(),
-                    this.configuration.getBlockReadRetryCount(),
-                    aggregatingMetrics,
-                    indexCache,
-                    openStreamInformation);
-            blockStore.add(blockKey, block);
-          }
-        });
   }
 
-  /** cleans data from memory */
-  public void cleanUp() {
-    blockStore.cleanUp();
+  /**
+   * Groups sequential block indexes into separate lists, ensuring each group doesn't exceed the
+   * maximum block count.
+   *
+   * @param blockIndexes an ordered list of block indexes
+   * @return a list of lists where each inner list contains sequential block indexes within size
+   *     limits
+   * @see RangeOptimiser#optimizeReads(List, long)
+   */
+  private List<List<Integer>> splitReads(List<Integer> blockIndexes) {
+    return rangeOptimiser.optimizeReads(blockIndexes, configuration.getReadBufferSize());
   }
 
-  private long getLastObjectByte() {
-    return this.metadata.getContentLength() - 1;
+  /**
+   * Detects sequential read pattern and finds the generation of the block
+   *
+   * @param pos position of the read
+   * @param readMode whether this ask corresponds to a sync or async read
+   * @return generation of the block
+   */
+  private long getGeneration(long pos, ReadMode readMode) {
+    // Generation is zero for ASYNC reads or first block of the object
+    if (readMode == ReadMode.ASYNC || pos < configuration.getReadBufferSize()) return 0;
+
+    Optional<Block> previousBlock = blockStore.getBlock(pos - 1);
+    return previousBlock.map(block -> block.getGeneration() + 1).orElse(0L);
   }
 
   private long truncatePos(long pos) {
@@ -244,9 +204,73 @@ public class BlockManager implements Closeable {
     return Math.min(pos, getLastObjectByte());
   }
 
+  private boolean isRangeAvailable(long pos, long endPos) {
+    // measure is true, since this is the first check if block exist or not
+    List<Integer> missingBlockIndexes = blockStore.getMissingBlockIndexesInRange(pos, endPos, true);
+    return missingBlockIndexes.isEmpty();
+  }
+
+  private long getLastObjectByte() {
+    return this.metadata.getContentLength() - 1;
+  }
+
+  /**
+   * Calculates the {@link Range} for a given block index within the S3 object.
+   *
+   * <p>The start of the range is calculated as {@code blockIndex * readBufferSize}. The end of the
+   * range is the smaller of:
+   *
+   * <ul>
+   *   <li>The last byte of the block: {@code ((blockIndex + 1) * readBufferSize) - 1}
+   *   <li>The last byte of the S3 object: {@code getLastObjectByte()}
+   * </ul>
+   *
+   * <p>This ensures that the returned range does not exceed the actual size of the object.
+   *
+   * @param blockIndex the index of the block for which the byte range is being calculated
+   * @return a {@link Range} representing the byte range [start, end] for the specified block
+   */
+  private Range getBlockIndexRange(int blockIndex) {
+    long start = blockIndex * configuration.getReadBufferSize();
+    long end = Math.min(start + configuration.getReadBufferSize() - 1, getLastObjectByte());
+    return new Range(start, end);
+  }
+
+  /**
+   * Retrieves the {@link Block} containing the given position, if it exists in the block store.
+   *
+   * @param pos the byte position within the object to look up
+   * @return an {@link Optional} containing the {@link Block} if present; otherwise, {@link
+   *     Optional#empty()}
+   */
+  public synchronized Optional<Block> getBlock(long pos) {
+    return this.blockStore.getBlock(pos);
+  }
+
+  /**
+   * Removes the specified {@link Block}s from the block store.
+   *
+   * @param blocks the list of {@link Block}s to remove
+   */
+  private synchronized void removeBlocks(final List<Block> blocks) {
+    blocks.forEach(blockStore::remove);
+  }
+
+  /**
+   * Checks whether the {@link BlockStore} currently holds any blocks.
+   *
+   * @return {@code true} if the block store is empty; {@code false} otherwise
+   */
+  public boolean isBlockStoreEmpty() {
+    return this.blockStore.isEmpty();
+  }
+
+  /** cleans data from memory */
+  public void cleanUp() {
+    this.blockStore.cleanUp();
+  }
+
   /** Closes the {@link BlockManager} and frees up all resources it holds */
   @Override
-  public void close() {
-    blockStore.close();
-  }
+  public void close() {}
 }
