--- input-stream/src/main/java/software/amazon/s3/analyticsaccelerator/io/physical/data/DataBlockStore.java
@@ -1,205 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package software.amazon.s3.analyticsaccelerator.io.physical.data;
-
-import java.io.Closeable;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import lombok.NonNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import software.amazon.s3.analyticsaccelerator.common.Metrics;
-import software.amazon.s3.analyticsaccelerator.common.Preconditions;
-import software.amazon.s3.analyticsaccelerator.io.physical.PhysicalIOConfiguration;
-import software.amazon.s3.analyticsaccelerator.util.BlockKey;
-import software.amazon.s3.analyticsaccelerator.util.MetricKey;
-
-/**
- * A container that manages a collection of {@link DataBlock} instances. Each {@code DataBlock}
- * corresponds to a fixed-size chunk of data based on the configured block size. This class provides
- * methods to retrieve, add, and track missing blocks within a specified data range.
- */
-public class DataBlockStore implements Closeable {
-
-  private static final Logger LOG = LoggerFactory.getLogger(DataBlockStore.class);
-
-  private final BlobStoreIndexCache indexCache;
-  private final Metrics aggregatingMetrics;
-  private final PhysicalIOConfiguration configuration;
-  // It is safe to use Integer as key since maximum single file size is 5TB in S3
-  // and if we assume that block size will be 8KB, total number of blocks is within range
-  // 5 TB / 8 KB = (5 * 1024^4) / 8192 ≈ 671,088,640 blocks
-  // Max int value = 2,147,483,647
-  private final Map<Integer, DataBlock> blocks;
-
-  /**
-   * Creates a new {@link DataBlockStore} with the specified configuration.
-   *
-   * @param configuration the {@link PhysicalIOConfiguration} used to define block size and other
-   *     I/O settings
-   * @param indexCache blobstore index cache
-   * @param aggregatingMetrics blobstore metrics
-   */
-  public DataBlockStore(
-      @NonNull BlobStoreIndexCache indexCache,
-      @NonNull Metrics aggregatingMetrics,
-      @NonNull PhysicalIOConfiguration configuration) {
-    this.indexCache = indexCache;
-    this.aggregatingMetrics = aggregatingMetrics;
-    this.configuration = configuration;
-    blocks = new ConcurrentHashMap<>();
-  }
-
-  /**
-   * Retrieves the {@link DataBlock} containing the byte at the specified position, if it exists.
-   *
-   * @param pos the byte offset to locate
-   * @return an {@link Optional} containing the {@code DataBlock} if found, or empty if not present
-   */
-  public Optional<DataBlock> getBlock(long pos) {
-    Preconditions.checkArgument(0 <= pos, "`pos` must not be negative");
-    return getBlockByIndex(getPositionIndex(pos));
-  }
-
-  /**
-   * Retrieves the {@link DataBlock} at the specified index from the block store.
-   *
-   * @param index the index of the block to retrieve
-   * @return an {@link Optional} containing the {@link DataBlock} if present; otherwise, an empty
-   *     {@link Optional}
-   */
-  public Optional<DataBlock> getBlockByIndex(int index) {
-    Preconditions.checkArgument(0 <= index, "`index` must not be negative");
-    return Optional.ofNullable(blocks.get(index));
-  }
-
-  /**
-   * Adds a new {@link DataBlock} to the store if a block at the corresponding index doesn't already
-   * exist.
-   *
-   * @param block the {@code DataBlock} to add
-   */
-  public void add(DataBlock block) {
-    this.blocks.putIfAbsent(getBlockIndex(block), block);
-  }
-
-  /**
-   * Removes the specified {@link DataBlock} from the store and updates memory usage metrics.
-   *
-   * @param block the {@code DataBlock} to remove
-   */
-  public void remove(DataBlock block) {
-    int blockIndex = getBlockIndex(block);
-    if (blocks.remove(blockIndex) != null) {
-      aggregatingMetrics.reduce(MetricKey.MEMORY_USAGE, block.getBlockKey().getRange().getLength());
-    }
-  }
-
-  /**
-   * Returns the list of block indexes that are missing for the given byte range.
-   *
-   * @param startPos the starting byte position (inclusive)
-   * @param endPos the ending byte position (inclusive)
-   * @param measure whether to measure cache hits and misses. If true, metrics will be updated.
-   * @return a list of missing block indexes within the specified range
-   */
-  public List<Integer> getMissingBlockIndexesInRange(long startPos, long endPos, boolean measure) {
-    return getMissingBlockIndexesInRange(
-        getPositionIndex(startPos), getPositionIndex(endPos), measure);
-  }
-
-  // TODO Consider using Range, otherwise add Preconditions to check start and end indexes
-  private List<Integer> getMissingBlockIndexesInRange(
-      int startIndex, int endIndex, boolean measure) {
-    List<Integer> missingBlockIndexes = new ArrayList<>();
-
-    for (int i = startIndex; i <= endIndex; i++) {
-      if (!blocks.containsKey(i)) {
-        missingBlockIndexes.add(i);
-
-        if (measure) aggregatingMetrics.add(MetricKey.CACHE_MISS, 1L);
-      } else {
-        if (measure) aggregatingMetrics.add(MetricKey.CACHE_HIT, 1L);
-      }
-    }
-    return missingBlockIndexes;
-  }
-
-  /**
-   * Cleans data from memory by removing blocks that are no longer needed. This method iterates
-   * through all blocks in memory and removes those that: 1. Have their data loaded AND 2. Are not
-   * present in the index cache For each removed block, the method: - Removes the block from the
-   * internal block store - Updates memory usage metrics
-   */
-  public void cleanUp() {
-    Iterator<Map.Entry<Integer, DataBlock>> iterator = blocks.entrySet().iterator();
-    while (iterator.hasNext()) {
-      Map.Entry<Integer, DataBlock> entry = iterator.next();
-      DataBlock block = entry.getValue();
-      BlockKey blockKey = block.getBlockKey();
-      if (block.isDataReady() && !indexCache.contains(blockKey)) {
-        try {
-          iterator.remove();
-          aggregatingMetrics.reduce(MetricKey.MEMORY_USAGE, blockKey.getRange().getLength());
-          LOG.debug(
-              "Removed block with key {}-{}-{} from block store during cleanup",
-              blockKey.getObjectKey().getS3URI(),
-              blockKey.getRange().getStart(),
-              blockKey.getRange().getEnd());
-        } catch (Exception e) {
-          LOG.error("Error in removing block {}", e.getMessage());
-        }
-      }
-    }
-  }
-
-  private int getBlockIndex(DataBlock block) {
-    return getPositionIndex(block.getBlockKey().getRange().getStart());
-  }
-
-  private int getPositionIndex(long pos) {
-    return (int) (pos / this.configuration.getReadBufferSize());
-  }
-
-  /**
-   * Closes all {@link DataBlock} instances in the store and clears the internal map. This should be
-   * called to release any underlying resources or memory.
-   *
-   * @throws IOException if an I/O error occurs during block closure
-   */
-  @Override
-  public void close() throws IOException {
-    // TODO Memory Manager
-    for (DataBlock block : blocks.values()) {
-      block.close();
-    }
-    blocks.clear();
-  }
-
-  /**
-   * Returns true if blockstore is empty
-   *
-   * @return true if blockstore is empty
-   */
-  public boolean isEmpty() {
-    return this.blocks.isEmpty();
-  }
-}
